feat: modified difficulty validation

+++ src/consensus/params.h
@@ -93,7 +93,9 @@ struct Params {
     /** Staking parameters */
     int nStakingStartHeight;
     uint32_t freeTxMaxSizeInBlock;
-    double freeTxDifficultyCoefficient;
+    // freeTxDifficultyCoefficient set to N would mean that the difficulty of a block filled with free transactions
+    // may be no lesser than (1 - 1 / N) * base_difficulty
+    uint32_t freeTxDifficultyCoefficient;
     std::array<size_t, stakingParams::NUM_STAKING_PERIODS> stakingPeriod;
     std::array<double, stakingParams::NUM_STAKING_PERIODS> stakingRewardPercentage;
 };
